requirements.txt
idna==2.10
chardet
httpx
inscriptis
PyYAML
html2text